Add delay before checking if pumping finished with an error
W/o the delay it might happen that an error (e.g. following error) did
not yet propagate from the EPOS controller to the SDK. Therefore the
SiLA Command would finish successfully while the drive was actually in
fault state and the desired volume had not been actually dosed.

Fixes #1376